[#102] Refactor UI to enforce one dataset per visualisation
