+++ client/src/utilities/chart.js
@@ -1,14 +1,12 @@
 export function getChartData(visualisation, datasets) {
-  const datasetIDX = visualisation.sourceDatasetX;
-  const datasetIDY = visualisation.sourceDatasetY;
-  const datasetX = datasets[datasetIDX];
-  const datasetY = datasets[datasetIDY];
+  const datasetID = visualisation.sourceDatasetX;
+  const dataset = datasets[datasetID];
   const columnIndexX = visualisation.datasetColumnX;
   const columnIndexY = visualisation.datasetColumnY;
-  const nameDataX = datasetX.columns[visualisation.datasetNameColumnX];
-  const dataX = datasetX.columns[columnIndexX].values;
-  const dataY = datasetY ? datasetY.columns[columnIndexY].values : null;
-  const dataValues = [];
+  const nameDataX = dataset.columns[visualisation.datasetNameColumnX];
+  const dataX = dataset.columns[columnIndexX].values;
+  const dataY = columnIndexY !== null ? dataset.columns[columnIndexY].values : null;
+  let dataValues = [];
   let output = [];
 
   switch (visualisation.visualisationType) {
@@ -16,14 +14,14 @@ export function getChartData(visualisation, datasets) {
     case 'line':
     case 'area':
 
-      dataX.forEach((entry, index) => {
+      dataValues = dataX.map((entry, index) => {
         let key = index;
 
         if (nameDataX && visualisation.visualisationType === 'bar') {
           key = nameDataX.values[index];
         }
 
-        dataValues.push({
+        return ({
           x: key,
           y: parseInt(entry, 10),
         });
@@ -33,36 +31,40 @@ export function getChartData(visualisation, datasets) {
         name: 'series1',
         values: dataValues,
       });
+
       break;
 
     case 'pie':
     case 'donut':
 
-      dataX.forEach((entry, index) => {
+
+      output = dataX.map((entry, index) => {
         const key = nameDataX ? nameDataX.values[index] : index;
 
-        dataValues.push({
+        return ({
           label: key,
           value: parseInt(entry, 10),
         });
       });
 
-      output = dataValues;
       break;
 
     case 'scatter':
 
-      dataX.forEach((entry, index) => {
-        dataValues.push({
+      dataValues = dataX.map((entry, index) => {
+        const item = {
           x: parseInt(entry, 10),
           y: parseInt(dataY[index], 10),
-        });
+        };
+
+        return item;
       });
 
       output.push({
         name: 'series1',
         values: dataValues,
       });
+
       break;
 
     default:
